Can now CRUD file

+++ files-lib.js
@@ -1,54 +1,76 @@
 const process = require('process')
 const fs = require('fs')
 const path = require('path')
+const formidable = require('formidable')
+const util = require('util')
+const copyFile = util.promisify(fs.copyFile)
+const mime = require('mime')
 
 const BASE_FOLDER = process.env.FOLDER
 
-function create(message, done) {
-    console.log("creating new file", message)
-    done();
+function create(req, done) {
+    var form = new formidable.IncomingForm()
+
+    form.parse(req, (err, fields, files) => {
+        if(err) throw err
+
+        Object.keys(files).map( key => {
+
+            let file = files[key]
+            let destination = BASE_FOLDER+file.name
+
+            // 1. copy file
+            // copyFile(file.path, destination)
+
+            // 2. fs writefile     error. cannot get fileContent as binary
+            // fs.writeFile(destination, file, 'binary', (err) => {
+            //     if(err) throw err
+            // })
+
+            // 3. rename file
+            fs.rename(file.path, destination, (err) => {
+                if(err) throw err
+                let result = {
+                    publicKey: file.name,
+                    privateKey: file.name
+                }
+                done(result)
+            })
+
+            // 4. base64
+            //  Try to convert file to base64
+        })
+    })
+
 }
 
 function download(publicKey, done) {
-    var filename = 'sample-image.jpg' // TODO: base from publickKey
-    var fileLocation = BASE_FOLDER+filename
-    var file
-    var error
+    var fileLocation = BASE_FOLDER+publicKey
     try{
-        file = fs.readFileSync(fileLocation)   // add slash between FOLDER and filename if FOLDER value has no slash ending
+        if(fs.existsSync(fileLocation)) {
+            file = fs.readFileSync(fileLocation)   // add slash between FOLDER and filename if FOLDER value has no slash ending
+            let mimetype = mime.lookup(fileLocation)
+            done(mimetype, file)
+        } else {
+            throw("Error: file does not exist")
+        }
     } catch(err) {
-        error = `Cannot read file ${fileLocation}`
+        throw(err)
     }
 
-    done(error, file)
 }
 
 function remove(privateKey, done) {
-    console.log('privateKey: ', privateKey)
-    var filename = 'sample-image.jpg'
-    var fileLocation = BASE_FOLDER+filename
+    var fileName = privateKey
+    var fileLocation = BASE_FOLDER+fileName
 
-    var file
-    var error
-    try {
-        // if(fs.existsSync(fileLocation)) {
-        //     console.log('file exist')
-            fs.unlink(fileLocation, (err) => {
-                if(err) error = err
-                // file successfully deleted. do nothing
-                console.log('done with error: ', error)
-                done(error)
-            })
-
-        // } else {
-        //     console.log("file does not exist")
-        //     error = `Error: Could not find file ${fileLocation}`
-        //     done(error)
-        // }
-
-    } catch (error) {
-        error = err
-        done(error)
+    if(fs.existsSync(fileLocation)) {
+        fs.unlink(fileLocation, (err) => {
+            if(err) throw err
+            done()          // file successfully deleted. do nothing
+        })
+    } else {
+        throw `Error: Could not fine file ${fileName}`
     }
 }
